Re-construct tests
