+++ pyroxide/src/codegen.rs
@@ -85,50 +85,3 @@ pub fn generate_from_wit(interfaces: &[wit_parser::Interface]) -> Result<String>
     }
     Ok(format(quote! { #(#tt)* }))
 }
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn generate_from_test_wit() {
-        let project_root = std::path::Path::new(env!("CARGO_MANIFEST_DIR"))
-            .parent()
-            .unwrap();
-        let interfaces = crate::wit::parse(&project_root.join("test.wit")).unwrap();
-        let tt = generate_from_wit(&interfaces).unwrap();
-        insta::assert_snapshot!(tt, @r###"
-        pub mod example {
-            use pyo3::{prelude::*, types::PyString};
-            pub fn a1<'py>(py: Python<'py>) -> PyResult<()> {
-                let _ = py.import("example")?.getattr("a1")?.call((), None)?;
-                Ok(())
-            }
-            pub fn a2<'py>(py: Python<'py>, x: i64) -> PyResult<()> {
-                let _ = py.import("example")?.getattr("a2")?.call((x,), None)?;
-                Ok(())
-            }
-            pub fn a3<'py>(py: Python<'py>, y: &str, z: f64) -> PyResult<()> {
-                let _ = py.import("example")?.getattr("a3")?.call((y, z), None)?;
-                Ok(())
-            }
-            pub fn a4<'py>(py: Python<'py>) -> PyResult<i64> {
-                let result = py.import("example")?.getattr("a4")?.call((), None)?;
-                Ok(result.extract()?)
-            }
-            pub fn a5<'py>(py: Python<'py>, x: i64) -> PyResult<&'py PyString> {
-                let result = py.import("example")?.getattr("a5")?.call((x,), None)?;
-                Ok(result.extract()?)
-            }
-            pub fn a6<'py>(py: Python<'py>) -> PyResult<(i64, &'py PyString)> {
-                let result = py.import("example")?.getattr("a6")?.call((), None)?;
-                Ok(result.extract()?)
-            }
-            pub fn a7<'py>(py: Python<'py>, x: i64) -> PyResult<(i64, &'py PyString, f64)> {
-                let result = py.import("example")?.getattr("a7")?.call((x,), None)?;
-                Ok(result.extract()?)
-            }
-        }
-        "###);
-    }
-}
